fix dep error by moving StoryAnalytics to src/analytics
