+++ extensions/amp-story/1.0/amp-story-page-attachment.js
@@ -28,9 +28,10 @@ import {
 } from './utils';
 import {Layout} from '../../../src/layout';
 import {Services} from '../../../src/services';
-import {StoryAnalyticsEvent, getAnalyticsService} from './story-analytics';
+import {StoryAnalyticsEvent} from '../../../src/analytics';
 import {closest, isAmpElement} from '../../../src/dom';
 import {dev} from '../../../src/log';
+import {getAnalyticsService} from './story-analytics';
 import {getState} from '../../../src/history';
 import {htmlFor} from '../../../src/static-template';
 import {listen} from '../../../src/event-helper';
