+++ src/analytics.js
@@ -31,3 +31,15 @@ export function triggerAnalyticsEvent(target, eventType, opt_vars) {
     analytics.triggerEventForTarget(target, eventType, opt_vars);
   });
 }
+
+/** @enum {string} */
+export const StoryAnalyticsEvent = {
+  BOOKEND_ENTER: 'story-bookend-enter',
+  BOOKEND_EXIT: 'story-bookend-exit',
+  LAST_PAGE_VISIBLE: 'story-last-page-visible',
+  PAGE_ATTACHMENT_ENTER: 'story-page-attachment-enter',
+  PAGE_ATTACHMENT_EXIT: 'story-page-attachment-exit',
+  PAGE_VISIBLE: 'story-page-visible',
+  STORY_MUTED: 'story-audio-muted',
+  STORY_UNMUTED: 'story-audio-unmuted',
+};
